requirements.txt
streamlit
opencv-python
numpy
pillow
pandas
face-recognition
pytesseract
python-dotenv
pydub
scikit-learn
qrcode